moved d3 draw functionality into ultils
