+++ client/components/utilities/drawNetworkGraph.js
